[feature] possible to extract computing and detection to other device
The raspberry pi sends data to another device that displays and hosts the website. For this purpose a stronger device is likely to be used. If the website usage is enabled the pi extracts the detection and computation to the other device, because it should be more powerfull. It is not possible yet to seperate computation and detection from website hosting.